+++ src/integrate2.cpp
@@ -36,7 +36,7 @@ struct integrate_data {
   component *cS;
   complex<double> *ph;
   complex<double> *fvals;
-  int *offsets;
+  ptrdiff_t *offsets;
   int ninveps;
   component inveps_cs[3];
   direction inveps_ds[3];
@@ -59,19 +59,19 @@ static void integrate_chunkloop(fields_chunk *fc, int ichunk, component cgrid,
 {
   (void) ichunk; // unused
   integrate_data *data = (integrate_data *) data_;
-  int *off = data->offsets;
+  ptrdiff_t *off = data->offsets;
   component *cS = data->cS;
   complex<double> *fvals = data->fvals, *ph = data->ph;
   complex<long double> sum = 0.0;
   double maxabs = 0;
   const component *iecs = data->inveps_cs;
   const direction *ieds = data->inveps_ds;
-  int ieos[6];
+  ptrdiff_t ieos[6];
   const component *imcs = data->invmu_cs;
   const direction *imds = data->invmu_ds;
   int num_fvals1 = data->num_fvals;
   int num_fvals2 = data->num_fvals2;
-  int imos[6];
+  ptrdiff_t imos[6];
   const fields_chunk *fc2 = data->fields2->chunks[ichunk];
 
   for (int i = 0; i < num_fvals1; ++i) {
@@ -176,7 +176,7 @@ static void integrate_chunkloop(fields_chunk *fc, int ichunk, component cgrid,
       }
     }
 
-    complex<double> integrand = 
+    complex<double> integrand =
       data->integrand(fvals, loc, data->integrand_data_);
     maxabs = max(maxabs, abs(integrand));
     sum += integrand * IVEC_LOOP_WEIGHT(s0, s1, e0, e1, dV0 + dV1 * loop_i2);
@@ -189,7 +189,7 @@ static void integrate_chunkloop(fields_chunk *fc, int ichunk, component cgrid,
 complex<double> fields::integrate2(const fields &fields2,
 				   int num_fvals1,
 				   const component *components1,
-				   int num_fvals2, 
+				   int num_fvals2,
 				   const component *components2,
 				   field_function integrand,
 				   void *integrand_data_,
@@ -203,9 +203,9 @@ complex<double> fields::integrate2(const fields &fields2,
     return integrate(num_fvals1, components1, integrand, integrand_data_,
 		     where, maxabs);
   if (num_fvals1 == 0)
-    return const_cast<fields&>(fields2).integrate(num_fvals2, components2, 
+    return const_cast<fields&>(fields2).integrate(num_fvals2, components2,
 			     integrand, integrand_data_, where, maxabs);
-  
+
   // check if components are all on the same grid:
   bool same_grid = true;
   for (int i = 1; i < num_fvals1; ++i)
@@ -245,14 +245,14 @@ complex<double> fields::integrate2(const fields &fields2,
     if (components1[i] == Dielectric) { needs_dielectric = true; break; }
   if (!needs_dielectric) for (int i = 0; i < num_fvals2; ++i)
     if (components2[i] == Dielectric) { needs_dielectric = true; break; }
-  if (needs_dielectric) 
+  if (needs_dielectric)
     FOR_ELECTRIC_COMPONENTS(c) if (gv.has_field(c)) {
       if (data.ninveps == 3) abort("more than 3 field components??");
       data.inveps_cs[data.ninveps] = c;
       data.inveps_ds[data.ninveps] = component_direction(c);
       ++data.ninveps;
     }
-  
+
   /* compute inverse-mu directions for computing Permeability fields */
   data.ninvmu = 0;
   bool needs_permeability = false;
@@ -260,15 +260,15 @@ complex<double> fields::integrate2(const fields &fields2,
     if (components1[i] == Permeability) { needs_permeability = true; break; }
   if (!needs_permeability) for (int i = 0; i < num_fvals2; ++i)
     if (components2[i] == Permeability) { needs_permeability = true; break; }
-  if (needs_permeability) 
+  if (needs_permeability)
     FOR_MAGNETIC_COMPONENTS(c) if (gv.has_field(c)) {
       if (data.ninvmu == 3) abort("more than 3 field components??");
       data.invmu_cs[data.ninvmu] = c;
       data.invmu_ds[data.ninvmu] = component_direction(c);
       ++data.ninvmu;
     }
-  
-  data.offsets = new int[2 * (num_fvals1 + num_fvals2)];
+
+  data.offsets = new ptrdiff_t[2 * (num_fvals1 + num_fvals2)];
   for (int i = 0; i < 2 * (num_fvals1 + num_fvals2); ++i)
     data.offsets[i] = 0;
 
@@ -286,8 +286,8 @@ complex<double> fields::integrate2(const fields &fields2,
   return complex<double>(real(data.sum), imag(data.sum));
 }
 
-typedef struct { 
-  field_rfunction integrand; void *integrand_data; 
+typedef struct {
+  field_rfunction integrand; void *integrand_data;
 } rfun_wrap_data;
 static complex<double> rfun_wrap(const complex<double> *fields,
 				 const vec &loc, void *data_) {
@@ -306,7 +306,7 @@ double fields::integrate2(const fields &fields2,
   rfun_wrap_data data;
   data.integrand = integrand;
   data.integrand_data = integrand_data_;
-  return real(integrate2(fields2, num_fvals1, components1, 
+  return real(integrate2(fields2, num_fvals1, components1,
 			 num_fvals2, components2,
 			 rfun_wrap,
 			 &data, where, maxabs));
